+++ libs/common/src/lib/interfaces/portfolio-position.interface.ts
@@ -1,6 +1,7 @@
 import { MarketState } from '@ghostfolio/api/services/interfaces/interfaces';
 import { AssetClass, AssetSubClass, DataSource } from '@prisma/client';
 
+import { Market } from '../types';
 import { Country } from './country.interface';
 import { Sector } from './sector.interface';
 
@@ -19,6 +20,7 @@ export interface PortfolioPosition {
   marketChange?: number;
   marketChangePercent?: number;
   marketPrice: number;
+  markets?: { [key in Market]: number };
   marketState: MarketState;
   name: string;
   netPerformance: number;