+++ libs/common/src/lib/types/market.type.ts
@@ -1,0 +1,1 @@
+export type Market = 'developedMarkets' | 'emergingMarkets' | 'otherMarkets';